+++ packages/mbfactory/src/commands/push-diff/push-diff.unit.test.ts
@@ -5,6 +5,7 @@ jest.mock('../../get-all-plugin-data/get-all-plugin-data', () => ({
 
 import { Arguments } from 'yargs'
 import { callHandler } from '../../call-handler/call-handler'
+import { ExitCodes } from '../../lib/exit-codes'
 import { logger } from '../../lib/logger'
 import { PushDiffArgv, pushDiffCommand } from './push-diff'
 
@@ -38,6 +39,7 @@ describe('The push-diff command should', () => {
       '- file not found: ./src/commands/push-diff/__fixtures__/xxx.ts'
     )
     expect(callHandler).toHaveBeenCalledTimes(0)
+    expect(process.exitCode).toEqual(ExitCodes.FileNotFound)
   })
 
   test('report nothing to process when file exists', async () => {
@@ -47,5 +49,6 @@ describe('The push-diff command should', () => {
       '- nothing to process in file: ./src/commands/push-diff/__fixtures__/dummy-file.ts'
     )
     expect(callHandler).toHaveBeenCalledTimes(0)
+    expect(process.exitCode).toEqual(ExitCodes.NothingToProcess)
   })
 })