+++ packages/plugin-contentful/src/report-content-type/report-entry-has-changes.ts
@@ -1,14 +1,16 @@
 import { ReportEntry } from './report-entries'
 
-export interface HasChanges {
+export interface ReportEntryHasChanges {
   reportEntry: ReportEntry
 }
 
 // Check the report for differences
-export const hasChanges = ({ reportEntry }: HasChanges): boolean => {
+export const reportEntryHasChanges = ({
+  reportEntry,
+}: ReportEntryHasChanges): boolean => {
   if (reportEntry.state !== 'equal') return true
   for (const subEntry of reportEntry.subEntries) {
-    const subResult = hasChanges({ reportEntry: subEntry })
+    const subResult = reportEntryHasChanges({ reportEntry: subEntry })
     // If a differnce is found, break the loop and return, otherwise continue
     // with the rest of the tree
     if (subResult) return true